Controller specification update & clean-up
This commit adds support for extra A2DP codecs. However, is seems that
client (tested only with Android) does not select "better" one. It might
be required to dig in some more into the Bluetooth specification draft.

As for now, the sink endpoint seems to be working (with some glitches).
It is possible to stream audio from the remote Bluetooth device and
capture it with the "bluealsa" PCM plugin. The audio can be forwarded
to some other ALSA device, e.g.: ecasound -i alsa,bluealsa -o alsa