Use GIO instead of buggy libdbus
Some say that the API of the libdbus is over-designed. It supposed to
make everyone happy, but unfortunately failed miserably. For a simple
usage one might not see all the internal entanglement caused by the
thread-safety... but if one tries to relay on this "feature", it is
not as simple as it looks.

For more info see this thread:
https://bugs.freedesktop.org/show_bug.cgi?id=857