PCM transfer callback for writing data
The PCM plugin might need some redesign. Currently it is not possible to
determine how much data can be transferred to the FIFO without blocking,
hence the pointer in the streaming mode does nothing.

We have to avoid (at any cost) creating special thread just for reading
out data from (or writing to) the PCM FIFO. Such an approach might allow
us to know the size and the head of our fake ring buffer beforehand, but
it will produce unnecessary data copying...