No replay is not an error
During termination Bluez sends various signals. One of those is a signal
responsible for clearing configuration. This signal will cause transport
shutdown. However, the transport interface is already removed from the
Bluez (we are not responding to this signal), so we won't receive any
response for the Release method call.