Use transport self-management for PCM release
Previous approach of keeping index for opened PCM FIFO and releasing them
based on the index lookup has one major flaw - it was not thread safe. It
could be, but the advantage of fast client lookup was not a bottleneck in
our design. The simplest solution (linear search) should suffice for now.