There is no point in returning NULL
Returning NULL value might generate errors on the D-Bus side. If the
destination does not listen for our reply (e.g. the interface might be
removed the seconds after the request), it will cause delivery error,
which will be logged as an error, however in our case it is not.

Also make sure, that if an error occurs, our registered objects will
be unregistered - proper cleanup won't bite.