Use thread cancellation
IO threads are designed to be very autonomous and efficient. There is no
place for poll timeouts or similar spinning checks. In such a design, it
is very hard to reliably terminate thread from the inside and from the
outside. If some system call is blocking (e.g. poll, read or write), it
is practically impossible to terminate it. One way of proceeding is to
send a signal to a thread and detect if blocking call was interrupted.
However, thread synchronization based on signals is a horrible idea.

Yet another possibility is to use thread cancellation mechanism. By many,
such an approach is also discouraged. However, in our case it seems to be
the cleanest solution possible. It is simple (if applicable) and provides
a great control over thread functionality.