Thread-based virtual hardware
After a long struggle with transfer callbacks and pointer positioning,
the virtual mmap mode will be used. It emulates hardware ring buffer
and readiness signaling. Since the timing is kept by the server, the
virtual hardware thread is very thin - it performs blocked reading and
writing, and moves forward our hardware pointer.

It seems, that this approach is compatible with all implementations
used to read and write PCM data. It was tested with mpv (both video
and audio), cmus, ecasound, arecord, aplay and audacity.