Postpone FIFO synchronization until IO thread
Synchronizing FIFO in the hw_params PCM callback results in blocking
this call until the other side is connected. For the capture mode this
might never happen - Bluetooth device might never start playing. This
blocking causes application which "tests" ALSA device to wait for the
playback. For now on, the blocking part (synchronization) is done in
the IO thread, which is transparent for the PCM plugin - all callback
functions are responsive.

This commit fixes audacity startup when capturing device is connected.