Global configuration structure
Passing configuration variable to every function is a labor of Sisyphus.
It is not practical, makes code maintenance harder and obfuscates it. If
BlueALSA was a library, such a burden could be explained, however, for a
standalone application it is not. Beside, using global variable reduces
the executable by omitting address dereferencing.