Use main thread ID for IO thread liveness test
Using the status to determine whether thread is active or not might be
misleading. What if the state will need to be changed. However, there
is a thread intrinsic property, which determines thread liveness - it
is thread ID. If used with care, it is great indicator.

Note:
This commit changes a little bit the IO thread management system - a
very fragile thingy. If thread termination is not working properly,
this commit should be the first thing to blame.