Separate substructures for A2DP, RFCOMM and SCO
In the current architecture of BlueALSA, transport is an abstract
responsible for handling events associated with given profile type.
In the future it might be wise to change the name from transport to
something more meaningful.

There is only one transport structure, which has to suffice for all
types. To accomplish that, this structure has to contain all required
attributes. For a better maintainability these attributes are grouped
using substructures.