Use event FD for pause state management
Previously, pthread condition variable was used to block IO thread while
in the pause state. Such an approach requires standard "synchronization"
code. However, we are not synchronizing anything, we only want to pause
IO thread - which in theory should be totally independent. So, since the
event file descriptor is available globally in the transport structure,
it seems to be more elegant to use it for this purpose.

Additionally, this commit fixes a deadlock caused by the client being
disconnected while the transport is still in the paused state.