Safeguard cancellation zone for IO threads
Thread cancellation is allowed only in the safety zone - poll and read
after the poll. It is easier to maintain strictly defined place for such
a potentially destructive functionality - resources can not be allocated
in such a zone. So, we have to make sure, that if we jump off this zone
by the usage of break or goto statement, the code which follows will not
be canceled - resources can be safely released.