Synchronize PCM FIFO opening
POSIX-compatible FIFO opening in the non-blocking mode might cause a
headache (blocking mode is a nightmare too). In order to synchronize
this process, we will use signal (message) sent via the communication
socket. The purpose of this signal is to indicate, that the reading
side is ready, so now we can open the writing side.

During certain conditions, this lack of synchronization causes the IO
thread associated with the SCO to fail during FIFO opening, and then
to block in the poll system call until some event occurs.

Fixes #22