BlueALSA audio player
It is possible to use BlueALSA server as a Bluetooth audio hub. However,
in order to actually play audio it is required to forward audio signal
from the BlueALSA capture PCM to some playback PCM. It is possible to do
this with any ALSA application, however such an approach has an overhead
of unnecessary buffer copying back and forth within the alsa-lib.

This small player allows to feed audio from the BT device directly into
the destination PCM device.