Real-time PCM delay
This code should provide reliable delay calculation for the PCM
transport. However, due to unknown reasons (I guess it is buggy, or
what's worse, not implemented) the delay reporting on the BlueZ side
is not working.

With the BlueZ v5.43, Jabra MOVE v2.3.0 is off by approximately 500
milliseconds. This delay has to be due to BlueZ internal buffering,
or Jabra firmware. BlueALSA internal delay is in the range of 30 ms.