Account for samples queued in the ring buffer
In the commit 81fde85 I've removed the ring buffer component from
the delay calculation. It has turned out (through HW PCM testing),
that this buffer also should be accounted. However, previous logic
was very naive - this buffer is not filled-up by default, it has
to be configured that way. So, right now the logic is correct.