Fix signal scaling for SCO
HSP does not specify how to mute speaker and/or microphone. It is up to
manufacturer, how 0 gain is interpreted. In my case (Jabra MOVE v2.3.0)
audio is not muted (it seems that +VGM is not supported at all). In such
a case, a software mute seems to be a reasonable idea.