Allow connecting multiple Bluetooth devices
It seems, that in order to allow connecting more than one Bluetooth
device with the A2DP profile, it is required to register many media
endpoints within the BlueZ. It might be some undocumented feature,
though...

So, let put aside any concerns about this "feature" portability.

This commit adds logic for on-demand A2DP endpoint registration within
the BlueZ stack. When the already registered endpoint is used, new one
will be registered. I have omitted the cleanup procedure (on purpose - I
do not want to introduce unnecessary spaghetti code), so if one connects
multitude of BT devices, lots of resources (in the BlueZ) might be
allocated, which will not be freed until bluealsa is terminated.
However, it is some theoretical edge-case.