IO multiplexing for SCO thread
This commit introduces IO multiplexing for read/write operations
performed in the SCO thread. From now on, every IO operation is
performed only when the corresponding file descriptor is ready for
particular IO operation. Also, it covers read/write error checking
in a more correct way.

For a simplification of a linear FIFO-like buffer usage, there is
a separate shared piece of code. Due to simplicity and the usage of
the preprocessor, this buffer wrapper should have essentially none
performance impact.