A2DP source support for apt-X vendor codec
Apt-X codec is a proprietary vendor codec for A2DP. It is not possible
to ship bluez-alsa in a binary format with this codec included. However,
if one has a license and the BT apt-X encoder library itself, it will
be possible to enable apt-X just with the --enable-aptx configuration
option (and maybe some cosmetic code modifications).