Share PCM FIFO file descriptor via domain socket
This commit simplifies PCM communication initialization procedure for
BlueALSA server and connected client. In the previous approach, we were
creating named FIFO, which was shared between server and client via the
communication socket. Such an approach required a lot of fancy tricks
used for opening both ends of the pipe. Right now, pipe is opened by the
pipe() call, so we do not have to worry about opening order. Also, we
can use poll() instantly, because file descriptor is available right
away.