headers: move public headers into a bluealsa subdirectory
to share the same include paths pre and post installation

had to also change the includes to the shared code in the source.

Other projects create symlinks at build time to solve this issue.  I
think it's nice being able to make that mapping when browsing the
source.

Also conditionally wrap public headers with "extern C" when using c++