RTP mark bit quirk workaround
It seems, that RTP fragmentation is a rocket science, and not everyone
gets it. A2DP specification states, that the media payload format is
defined in the RFC 3016 (obsoleted by the RFC 6416), and I quote:

The marker bit is set to 1 to indicate that the RTP packet contains
a complete audioMuxElement or the last fragment of an audioMuxElement.

This commit adds a workaround for cases where mark bit is set to zero
for all packets.

Fixes #130