Read notification events from separate connection
Reading events from the same connection which is used for sending
commands leads to message mixing (connection is not reliable). One
approach would be to send events as a OOB data, however UNIX domain
stream socket does not support such a feature. As a simple workaround
we are going to open dedicated connection to the bluealsa server just
for reading events.

Closes #134