Increase the maximal AAC bitrate to 320 kbit/s
The configuration structure supports up to 8.39 Mbit/s. However, using
the maximal possible value might be catastrophic. It is not possible for
Bluetooth to transmit data at the rate of 8 Mbit/s, so if A2DP source
does not select bitrate consciously, but blindly uses provided max
value (bluealsa does that :D), we will end up with a non-working setup.