Fix oFono SCO premature release
In order to properly handle SCO socket file descriptor obtained from
the oFono, it is required to omit our internal acquire/release actions
executed during event handling. Also, it is not possible not to read
data from the socket when there is no PCM client connected, because it
will be not possible to release SCO (we have to read from it in order
to detect hangup).

Fixes #172