Use LD_PRELOAD for patching dlopen
ALSA library does not provide any API for changing hard-coded directory
for software plug-ins. Since, we need to test our PCM plug-in BEFORE the
installation, there is no official way of doing it. However, since ALSA
library uses dlopen() for loading plug-ins, it is possible to patch the
behavior of this function via library preloading.