Improve bluealsa_poll_revents() callback
Make bluealsa behave more like the hw device when used with poll().
Ensure that calling poll() when the PCM is not in the RUNNING or
DRAINING states never locks up the application, and that the POLLERR
flag is set in such cases. Also, do not report POLLERR in PREPARED.

Fixes #192