Work around rounding error at volume 0 when setting ALSA volume
Due to rounding errors, a slightly negative value can be passed to
set_normalized_volume(), which will make the log10() call fail.
Actually, volume 0 is already failing because log10(0) is illegal.  So
let's fix this by implementing two corner cases: <=0 and >=100.

This commit is based on a fix for corner cases applied in MPD:
https://github.com/MusicPlayerDaemon/MPD/commit/bede564618e78d2a4eee5af329178be89ae34412
https://github.com/MusicPlayerDaemon/MPD/commit/6d48a5684a2a05176b5dc0107e557edffb7d0673

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>