Rewind LATM buffer on aacDecode error
If latm buffer is not rewinded on unsuccessful decoding, same bt frame 
data will be retried for decoding. If the frame is unable to decode latm
buffer will be allocated more memory to store new bt frames and will 
eventually cause OOM. Since retry on same frame is not going to cause
aac decode to success, it is best to just drop the frame and carry on.

Fixes #215