src/ba-transport.c: use correct mtu for SCO links
For SCO audio links, we *must* send data in chunks of the link mtu size to ensure correct timing. To use a constant value of 48 or 24 bytes is just wrong.

Obtain the link config *after* successful connection, because before then the mtu has not been negotiated with the remote device