Add ALSA mixer volume control support to aplay
Add a new command line option `-M` or `--mixer=NAME` to allow using
an ALSA mixer for volume control. For using the native volume control
the bluealsa server needs to be started with the `--a2dp-volume` command
line option, e.g.:

```
$ bluealsa -p a2dp-sink --a2dp-volume
$ bluealsa-aplay -D default -M Master --profile-a2dp XX:XX:XX:XX:XX:XX
```

Setting the volume works in both ways.

Note, that if the bluealsa server is started with `--a2dp-volume` and no
mixer control name is passed to bluealsa-aplay, the volume will not change
when the source changes the volume.

Signed-off-by: Jörg Krause <joerg.krause@embedded.rocks>