ba-transport.c: properly handle simultaneous a2dp acquire and release
A client that closes its a2dp pcm handle then immediately opens the same device can cause the daemon to start acquistion for the second open before release of the first has completed. Thus we need to add thread synchronisation to prevent corruption of the transport state.

This fix relies on an assumption that the IO thread holds the cleanup lock throughout the time that transport release in in progress