Dedicated dispatcher for incoming SCO links
SCO connection can be established either by AG or headset, where the
latter is less common. In order to support headset mode we have to be
able to accept incoming SCO links. It has turned out, that SCO socket
itself shall be used in a server (listen) and a client (connect) mode,
respectively for accepting incoming links and establishing new ones.

From now on, it should be possible to connect audio gateway like phone
to the bluealsa in the native HSP/HFP headset mode.