Proper management of ongoing SCO connection
Managing SCO connection requires a little bit of care.

As a first rule, do not terminate connection, if some other client is
using it. E.g. when SCO connection is used by PCM mic and PCM spk, and
one of these closes, then we have to keep going until the other one
closes too.

Secondly, we will allow an incoming connection to be established even
though no one is reading it. Such behavior will be consistent with the
A2DP connections, where we allow external source to connect with our
sink regardless of PCM client availability. Such a way of management
will allow PCM client to connect later and read incoming data.