Provide separate D-Bus path for sink and source
Previous approach was limited to symmetrical A2DP codecs (same settings
for sink and source in case of back-channel microphone connection). In
order to support FastStream codec, it is required to provide separate
D-Bus paths for sink and source, so different settings can be exported.

This commit breaks backward compatibility.