Handle D-Bus method calls in a separate thread
D-Bus communication is handled by a single thread - default-context
thread. When handling method call or e.g. signal we will block D-Bus
communication until we will return to the main event loop. There might
be a case where we need to call an external D-Bus service from within
our D-Bus handler which in return might needs to call us (e.g. calling
BlueZ for endpoint reconnection). In such case we will not handle
incoming call because BlueALSA is not in the main event loop.

In order to be able to handle the scenario depicted above we are going
to handle incoming message call in a separate thread - DBusConnection
is thread-safe. In such a way, we will dispatch method call handling
to another thread and promptly return to the main event loop.