Fix synchronization for PCM pause functionality
This commit adds thread synchronization logic when pausing/unpausing
to make sure the server does not pause while a transfer is in progress.
This is necessary to maintain the transaction timing.

Also, this commit fixes race condition during PCM file descriptors
closing in case of termination from the BlueALSA server end.

Co-authored-by: Arkadiusz Bokowy <arkadiusz.bokowy@gmail.com>