Do not notify BT device in case of software volume
With software volume control we will cut any ties with Bluetooth
device, so we can scale audio on our own. This will not apply to
BlueALSA running as a "sink", because Bluetooth source device is
not responsible for audio scaling.

This commit fixes volume double scaling with bluez >= 5.55, where
A2DP volume property was fixed for headsets without player profile
connected.