Allow XAPL response device name customization
It has turned out that not every device accepts BlueALSA as a device
name in the XAPL response. Such devices are, e.g.: Sony WF-1000XM3 and
WM-1000XM4. As a consequence, headset might not send us battery level
notifications.

In the Apple specification for Bluetooth Accessory Identification the
name is set as 'iPhone' and by the font type one might judge, that it
is a part of the specification (shall not be customized).

Co-authored-by: Alex Davis <alex14641@yahoo.com>