Synchronize PCM and transport IO thread readiness
When a client issues PCM open request, it might expect that when the
call returns the BlueALSA server will be ready to serve audio right
away. However, for A2DP Sources the transport acquisition is made
asynchronously and then the audio processing is also started in the
asynchronous manner. All this might contribute to initial lag in the
audio transfer. By adding synchronization between PCM open request
and an IO thread main event loop readiness, we will eliminate this
potential initial lag.