fix alsa hw params constraints
Current pcm plugin code results in very strange buffer and period size
selections if the application sets period size before buffer size. This
simple fix corrects that issue. There does not appear to be any clue in
the ALSA docs as to why the period size max needs to be so high, but
looking at other io plugin examples they all appear to set the max
period to at least 1MB and in testing this change always gives sensible
values for both period and bufffer sizes.