Split SCO IO thread into encoder and decoder
Currently only the mSBC codec is fully separated into an encoder and a
decoder IO thread. The CVSD is still handled in a single thread with a
simple multi-thread wrapper for thread signals forwarding. The reason
for that is that the CVSD seems not to work properly with USB-based BT
chips (tested with Intel chip) when reading and writing to the SCO link
is not synchronized.