Set initial PCM volume from ALSA configuration
Permits per-device, and per-user, setting of initial PCM volume
level and mute state from ALSA configuration parameters. Volume
level is specified as a percentage of maximum volume, and mute is
enabled by appending '-' to the volume level, or disabled by
appending '+'.

If no volume level is specified, the default is to leave the level
unchanged, If no '-' or '+' suffix is given, the default is to
leave the mute state unchanged.

For 2-channel PCMs the specified volume and mute state is applied
to both channels.