Improve ALSA PCM disconnection reporting
The previous approach required applications to check for ENODEV
from snd_pcm_recover() or snd_pcm_prepare(), and many do not do
this, resulting in an infinite loop of recovery attempts.

This change permits snd_pcm_writei() and snd_pcm_readi() to
return ENODEV by preventing the ioplug code from overwriting the
PCM state change. Most applications now see a BlueALSA PCM
disconnect in the same way as a USB device unplugged, and handle
it in the same way.