Stabilize bluealsa-aplay volume management
Rounding error in the conversion from dB to loudness causes the
bluetooth transport volume to be reduced each time a device
connects. This effect is particularly annoying with SCO transports
because of the narrow 15-point volume scale.

This simple change rounds the double generated by pow() to the
nearest integer instead of truncating it, which results in a
much more stable volume conversion.