Fix bluealsa-aplay when codec is changed
A change of codec by a connected device may result in a change in the
audio parameters, which in turn necessitates changing the hw_params of
the ALSA output device. For example if a SCO connection changes from
mSBC to CVSD the sampling rate changes from 16kHz to 8kHz. If left
running, the output device continues to play at 16kHz resulting in
double-speed playback and constant underruns.

Closes #522