ALSA Hook PCM for HFP-AG support
Many HFP-HF devices require the AG to initiate a call session
before enabling the audio stream; some for both microphone and
speaker, some for only the microphone. Since the requirements of
devices vary, it is best if BlueALSA server does not force a call
session in all cases; the AG user should be able to choose whether
to enable a call session on a case-by case basis.

This commit introduces a wrapper for the BlueALSA ALSA plugin that
sends the HFP call transfer and termination RFCOMM messages when
selected by the user.