Do not enable any BT profile by default
Rationale:

The BlueALSA service can be used in many different ways. There is no
default usage for it. If one wants to build Bluetooth speaker then sink
profiles shall be enabled. If one wants to use Bluetooth headset with a
host on which BlueALSA is running then one shall use source profiles.
Also, there is no default strategy for which sink/source profiles shall
be enabled. We've got A2DP, HSP and HFP.

Taking into account the above statement, it is reasonable to allow a
user to enable suitable set of BT profiles. However, providing a default
set of profiles might be confusing or not easy to use. The first case is
when we provide default profiles, but if user specifies a profile on a
command line, default set of profiles is not used. The latter case is
when we provide a syntax for enabling/disabling BT profiles. Then, in
case when source profiles are enabled by default but user wants only
sink ones, it will be required to manually disable default set, which
is IMHO a configuration burden.

So we have to admit that in case of BT profiles there is no sane
default. From now on BlueALSA will require to specify which profiles
should be enabled.