Fix RTP timestamp calculations for A2DP streams
The A2DP specification defines how the RTP timestamp shall be
calculated, but only for SBC and ATRAC codecs. According to the
specification the timestamp clock rate shall be equal to the audio
sampling rate. For MPEG12 and ACC the timestamp clock rate is
defined in RFC-2250 and RFC-3016 respectively. For these codecs
it shall be equal to 90kHz.

Additionally, this commit adds logic for checking incoming RTP
frames. It detects missing RTP packets based on sequence number.
Also, it detects missing PCM frames (samples per channel) based
on timestamp differences.