Allow systemd services arguments configuration
We like to have a way to pass the preferred arguments to be enabled
in systemd services at build time. This is useful to OSes that are
generated from sources, like Yocto.

To achieve that, we now introduce two configure options:
--with-systemdbluealsaargs: defaults to '-p a2dp-source -p a2dp-sink'
--with-systemdbluealsaaplayargs: defaults to empty string

They would be written to systemd services.

Due to the default values being used, there is no functional changes
before and after this patch.

Closes #565

Signed-off-by: Ming Liu <liu.ming50@gmail.com>