Apply rate plugin fix for alsa-lib 1.2.6 also
There is an error in the alsa-lib fix for the rate plugin that can
result in short writes (i.e. snd_pcm_writei() returns before the
requested buffer is completely written) if the "slave" plugin has
more free buffer space than the rate plugin.

The existing workaround for the original rate plugin bug also
avoids this error, so we enable the workaround code for current
alsa-lib releases too.

This commit also cleans up the workaround code.