Improve SCO socket transfer size selection and timing
Linux btusb module does not report the required message size
for SCO sockets, rather the "MTU" it reports is actually the
controller SCO buffer size. However, we can obtain the correct
size by reading the first incoming message. SCO packet transfers
begin, in both directions, as soon as the transport is acquired,
so we are guaranteed to have incoming packets even if the remote
microphone is not enabled (provided the adapter is routing
incoming SCO through the HCI).

This commit adjusts the outgoing message size, if necessary, after
the first incoming message has been received.

For mSBC streams the timing logic is also modified so that no USB
slots are missed if the client runs too slowly.