Fix SCO routing change for systemd non-root user
To send raw HCI commands application requires the CAP_NET_RAW
capability. In the current bluealsa.service.in file all capabilities
are denied by the PrivateUsers=true directive.

This patch permits only this one required capability, and in doing so
does not change the security exposure score of the unit. Now, it will be
possible to have SCO routing set by bluealsa for Broadcom adapters when
running as non-root user.

Closes #596