Mask all signals if thread cancellation is used
It seems that thread cancellation implementation in glibc 2.31 (and
maybe other versions as well) is flawed. If during the initial stage
of the cancellation procedure some other signal is delivered to the
thread (thread cancellation is implemented by delivering a signal to
canceled thread) and the signal handler is using POSIX cancellation
point function, we will end up in a deadlock.

To prevent such scenario, threads managed by thread cancellation must
have all signals masked.

All credits go to @borine for the solution to the thread cancellation
deadlock problem.