Enable oFono AG support with mSBC codec
BlueALSA refuses to acquire transport on client Open() call if the
codec is not set; and oFono will not create the codec connection
until the first acquisition. We avoid this stand-off by performing
a temporary acquisition request when the device connects. The codec
reported by the response to this request from oFono is then used
as the codec for subsequent client requests.