Improve HFP-HF to support gateways that use oFono
When creating a Codec connection, oFono, in AG mode, attempts to
open the SCO connection before acknowledging the HF response by
sending the final OK. This causes a BlueALSA HF to accept the SCO
socket before the codec is known.

This change sets the codec in the HF without waiting for that
final OK so that gateway devices based on oFono (e.g. Sailfish OS,
etc.) can use mSBC effectively with BlueALSA HF devices.