Fix SCO volume control for HF/HS profiles
A SCO transport always references the encoder thread as sco.spk_pcm and
the decoder thread as sco.mic_pcm. However, on a HF/HS node, the encoder
thread handles the microphone stream, and the decoder thread handles the
speaker stream. As a result the transport thread reference names are
misleading; and this has resulted in the RFCOMM VGM and VGS commands and
responses being applied to the wrong PCM on the HF/HS. This commit
corrects that error.