Fix transport stopping for terminated threads
In case when IO threads are already terminated, we can not send
cancel threads request to thread manager and wait for state change
of IO threads. In that case state transition from terminated to
stopping is not valid, so the wait function will not block, and
we will leave cancel threads request in the thread manager queue.
This can lead to unexpected IO threads termination in the nearest
future.

Fixes #585