Make transport acquisition an atomic operation
When both source and sink clients connect in quick succession, it
is possible that two client event handler threads will invoke
ba_transport_acquire() at the same time. Only one will call the
profile acquire() function, which is protected by the bt_fd_mtx
mutex. However that mutex is unlocked before the thread states are
set to BA_TRANSPORT_THREAD_STATE_IDLE; so there is no guarantee
that the thread which set the "acquired" flag will be the first to
reach the point where that flag is tested. It is therefore possible
that one thread will leave the acquisition operation with the
I/O thread states still set to BA_TRANSPORT_THREAD_STATE_TERMINATED.

This is fixed by adding a dedicated mutex to ensure that the
entire ba_transport_acquire() function is treated as a critical
section.