When changing SEP select one with the lowest index
This commit fixes issue with BlueZ < 5.67 where it was not possible to
register new A2DP endpoint with already registered application. In that
case newly exported endpoint is in fact not registered within BlueZ, so
it returns "Invalid arguments" in case someone wants to use it. Now, we
will use pre-registered endpoints before using new ones, so in the most
sample case everything should work fine even with BlueZ < 5.67.

Fixes #649