Improve ALSA PCM plugin drain functionality
Implement support for applications that use snd_pcm_abort() or
snd_pcm_nonblock(pcm, 2) to request that blocking calls are aborted
by a signal. Also apply a timeout to the local drain so that an
application can not be blocked indefinitely should the BlueALSA
server stop reading from the PCM FIFO.

Because the BlueALSA server does not have a non-blocking drain
operation, we implement non-blocking drain by omitting the server
drain request. This means that a small amount of audio at the end
of a stream may be lost. Most, possibly all, applications clear
the PCM non-block flag before calling snd_pcm_drain(), so this
incomplete drain is unlikely to impact any real use cases.