Do not block forever PCM client on control command
To prevent a client from being blocked forever waiting for a
control command response in case the server has failed, set a
time limit for the response to arrive. The default timeout is an
arbitrary 200ms for "Drop", "Pause", and "Resume", and 3000ms for
"Drain". Allow the application to set alternative timeouts as
required.