Fix D-Bus mock bus termination in our unit test
The g_test_dbus_down() waits for an event on the default main loop. Such
loop is run by this function, however, if the default main loop is
already run by the application, the event might be delivered prematurely
causing the g_test_dbus_down() to wait for a 30s safety timeout.