Pass A2DP configuration with NAME:CONFIG notation
In all places, the A2DP codec configuration is printed as NAME:CONFIG.
Also, when selecting configuration with ALSA plug-in, the notation is
just like that. However, when selecting codec with our CLI the config
needs to be passed as a separate argument. This might be confusing to
some. This commit unifies the notation of A2DP codec name with the
configuration in all places. However, it slightly breaks command line
syntax with previous releases of bluez-alsa.