Use basename implementation from glib-2.0
This is portable across various platforms and system C libraries e.g.
basename differs between glibc and musl, where glibc provides a GNU
implementation and POSIX version, and musl provides posix version only
and Newer version of musl have removed prototype for basename in string.h [1]
which caused build failures with gcc-14 on musl systems.

g_path_get_basename() is implemented in glib-2.0 and is portable across
these systems. It allocates a new string for the returned value which
should be free'd by user, however, all our usecases here are in main()
functions, so technically we are not going to leak any memory.

[1] https://git.musl-libc.org/cgit/musl/commit/?id=725e17ed6dff4d0cd22487bb64470881e86a92e7