aplay: Decrease the size of ALSA PCM buffer
It seems that safari on the iPhone cannot adjust A/V sync when
the audio latency exceeds about 400ms. So to keep the default
delay when using bluealsa-aplay below that critical value, change
the default ALSA hw params to lower values which are still
compatible with the majority of sound cards.