Do not install config directly into $alsaconfdir
Place the ALSA plugin configuration file into the $datadir tree,
and then create a symbolic link in the $alsaconfdir to complete the
installation. This follows the approach used by the alsa-plugins
repository for its plugins, and allows the local admin to disable
or enable the plugin simply by removing or re-creating the symbolic
link.

For compatibility with older ALSA releases this change is
conditional on ALSA >= 1.1.7